NewUpdating

+++ database/migrations/2021_05_24_025618_create_demarcheurs_table.php
@@ -18,7 +18,7 @@ class CreateDemarcheursTable extends Migration
             $table->string('nom');
             $table->string('prenom');
             $table->string('email');
-            $table->float('telephone');
+            $table->integer('telephone');
             $table->string('login');
             $table->string('password');
             $table->string('statut');
